MAINT: use float64 in annualize_compounding

+++ empyrical/utils.py
@@ -449,4 +449,4 @@ def annualize_compounding(stat, ann_factor=252):
     """
     Annualizes the given statistic based on `ann_factor`.
     """
-    return (stat + 1) ** ann_factor - 1
+    return np.float64(stat + 1) ** np.float64(ann_factor) - 1
